Adds BBLegacy, BBResult
BBLegacy is basicaly a copy of version 2.7.2 of this library.  Added to make sure those using the old versions of this class
be able to download this new version without worrying about anything breaking.

So if you call $bb->tournament_create, the new BinaryBeast::__call will try to execute it from BBLegacy

BBResult is a simple service result wrapper, that allows developers to flexibly refer to values from a result, so for exmaple
a developer could get the result code in any of these ways:
$result->result
$result['RESULT']
$result->RESULT
$result['result']
etc etc

Legacy service requests results are automatically wrapped